common.logic: move non-public implementation to common._logic
